ModPlug: Simplify a structure and remove macro check from it
Keep it for loader itself and for usage place only.

Note: ModPlug_Tell is a custom-added public function into my fork here:
https://github.com/WohlSoft/AudioCodecs/commit/950993938869fdc4850d6a374050add7168ffe47
@sezero, maybe you'll pick it up into your fork also?

Also add the `MODPLUG_HAS_TELL` macro into public header to make it easier recognize the existence of this call without making of extra unit tests for that.