music_id3tag.c: Drop Extended ID3v1 (TAG+)
Originally, this tag was created by a crazy Czech guy to oppose ID3v2. Unfortunately, this tag was never popular and no more exist in a wild. It MIGHT exist on very rare cases we'll never meet in present, except the case of digging a content of a 20 year old HDD.