music_mad.c: Avoid infinite loop caused by a bad MP3 file
I have an MP3 file that has no frames in a stream, but has a junk. This causing MAD to try read a first frame and fail. This fail wasn't caught on a first frame, therefore, a thing will suck in MAD_GetSome() because it thinks that there are any data to play would appear, but in fact - nope.