music_wav.c: Added workaround for MFAudio's WAV files
There are WAV files generated by MFAudio that has the LIST chunk with the chunk length longer than actually written at end of the file. So, verify, does LIST chunk is actually at end of the file. Do fail when nothing got read, or when read less bytes than the chunk length and it's not an end of the file.

#73